api/requirements.txt
flask-mysql
flask_sqlalchemy
flask-praetorian
flask_cors